+++ puma/plot_base.py
@@ -24,13 +24,13 @@ class PlotLineObject:  # pylint: disable=too-many-instance-attributes
     xmax : float, optional
         Maximum value of the x-axis, by default None
     colour : str, optional
-        colour of the object, by default None
+        Colour of the object, by default None
     label : str, optional
-        label of object, by default None
+        Label of object, by default None
     linestyle : str, optional
-        linestyle following numpy style, by default None
+        Linestyle following numpy style, by default None
     alpha : float, optional
-       Value for visibility of the plot lines, by default None
+        Value for visibility of the plot lines, by default None
     marker : str, optional
         Marker that is used in the plot. For example an x.
         By default None
@@ -59,7 +59,7 @@ class PlotLineObject:  # pylint: disable=too-many-instance-attributes
 # TODO: enable `kw_only` when switching to Python 3.10
 # @dataclass(kw_only=True)
 @dataclass
-class PlotObject:
+class PlotObject:  # pylint: disable=too-many-instance-attributes
     """Data base class defining properties of a plot object.
 
     Parameters
@@ -87,6 +87,8 @@ class PlotObject:
     y_scale : float, optional
         Scaling up the y axis, e.g. to fit the ATLAS Tag. Applied if ymax not defined,
         by default 1.3
+    logx: bool, optional
+        Set the log of x-axis, by default False
     logy : bool, optional
         Set log of y-axis of main panel, by default True
     xlabel : str, optional
@@ -153,6 +155,7 @@ class PlotObject:
     ymin_ratio_2: float = None
     ymax_ratio_2: float = None
     y_scale: float = 1.3
+    logx: bool = False
     logy: bool = True
     xlabel: str = None
     ylabel: str = None
@@ -192,7 +195,7 @@ class PlotObject:
         Raises
         ------
         ValueError
-            if n_ratio_panels not in [0, 1, 2]
+            If n_ratio_panels not in [0, 1, 2]
         """
         self.__check_figsize()
         allowed_n_ratio_panels = [0, 1, 2]
@@ -219,7 +222,7 @@ class PlotObject:
         Raises
         ------
         ValueError
-            if shape of `figsize` is not a tuple or list with length 2
+            If shape of `figsize` is not a tuple or list with length 2
         """
         if self.figsize is None:
             return
@@ -232,7 +235,7 @@ class PlotObject:
             )
 
 
-class PlotBase(PlotObject):
+class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
     """Base class for plotting"""
 
     def __init__(self, **kwargs) -> None:
@@ -241,7 +244,7 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         **kwargs : kwargs
-            kwargs from `plot_object`
+            Keyword arguments from `puma.PlotObject`
         """
         super().__init__(**kwargs)
         self.axis_top = None
@@ -258,16 +261,16 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         sub_plot_index : int, optional
-            indicates for the scenario with one ratio how large the upper and lower
+            Indicates for the scenario with one ratio how large the upper and lower
             panels are, by default 5
         """
         # TODO: switch to cases syntax in python 3.10
 
         if self.vertical_split:
             # split figure vertically instead of horizonally
-            if self.n_ratio_panels <= 1:
+            if self.n_ratio_panels >= 1:
                 logger.warning(
-                    "You set the number of ratio panels to %i"
+                    "You set the number of ratio panels to %i "
                     "but also set the vertical splitting to True. Therefore no ratio"
                     "panels are created.",
                     self.n_ratio_panels,
@@ -326,7 +329,7 @@ class PlotBase(PlotObject):
         same_height: bool = False,
         colour: str = "#920000",
         fontsize: int = 10,
-    ):
+    ):  # pylint: disable=too-many-arguments
         """Drawing working points in plot
 
         Parameters
@@ -342,9 +345,9 @@ class PlotBase(PlotObject):
             upper plot (0 is bottom, 1 is top). Must be the same
             order as the vlines_xvalues and the labels. By default None
         same_height : bool, optional
-            working point lines on same height, by default False
+            Working point lines on same height, by default False
         colour : str, optional
-            colour of the vertical line, by default "#920000"
+            Colour of the vertical line, by default "#920000"
         fontsize : int, optional
             Fontsize of the vertical line text. By default 10.
         """
@@ -390,28 +393,45 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         title : str, optional
-            title of top panel, if None using the value form the class variables,
+            Title of top panel, if None using the value form the class variables,
             by default None
         **kwargs : kwargs
-            kwargs passed to `matplotlib.axes.Axes.set_title()`
+            Keyword arguments passed to `matplotlib.axes.Axes.set_title()`
         """
         self.axis_top.set_title(self.title if title is None else title, **kwargs)
 
-    def set_logy(self, force: bool = False):
-        """Set log scale of y-axis of main panel.
+    def set_log(self, force: bool = False):
+        """Set log scale of the axes. For the y-axis, only the main panel is
+        set. For the x-axes (also from the ratio subpanels), all are changed.
 
         Parameters
         ----------
         force : bool, optional
-            forcing logy even if class variable is False, by default False
+            Forcing log even if class variable is False, by default False
         """
-        if not self.logy and not force:
-            return
-        if not self.logy:
-            logger.warning("Setting log of y-axis but `logy` flag was set to False.")
-        self.axis_top.set_yscale("log")
-        ymin, ymax = self.axis_top.get_ylim()
-        self.y_scale = ymin * ((ymax / ymin) ** self.y_scale) / ymax
+
+        if self.logx or force:
+            if not self.logx:
+                logger.warning(
+                    "Setting log of x-axis but `logx` flag was set to False."
+                )
+
+            # Set log scale for all plots
+            self.axis_top.set_xscale("log")
+            if self.axis_ratio_1:
+                self.axis_ratio_1.set_xscale("log")
+            if self.axis_ratio_2:
+                self.axis_ratio_2.set_xscale("log")
+
+        if self.logy or force:
+            if not self.logy:
+                logger.warning(
+                    "Setting log of y-axis but `logy` flag was set to False."
+                )
+
+            self.axis_top.set_yscale("log")
+            ymin, ymax = self.axis_top.get_ylim()
+            self.y_scale = ymin * ((ymax / ymin) ** self.y_scale) / ymax
 
     def set_y_lim(self):
         """Set limits of y-axis."""
@@ -455,9 +475,9 @@ class PlotBase(PlotObject):
         label : str, optional
             x-axis label, by default None
         align_right : bool, optional
-            alignment of y-axis label, by default True
+            Alignment of y-axis label, by default True
         **kwargs, kwargs
-            kwargs passed to `matplotlib.axes.Axes.set_ylabel()`
+            Keyword arguments passed to `matplotlib.axes.Axes.set_ylabel()`
         """
         label_options = {}
         if align_right:
@@ -485,7 +505,7 @@ class PlotBase(PlotObject):
         label : str, optional
             x-axis label, by default None
         **kwargs : kwargs
-            kwargs passed to `matplotlib.axes.Axes.set_xlabel()`
+            Keyword arguments passed to `matplotlib.axes.Axes.set_xlabel()`
         """
         xlabel_args = {
             "xlabel": self.xlabel if label is None else label,
@@ -507,10 +527,10 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         labelsize : int, optional
-            label size of x- and y- axis ticks, by default None
-            if None then using global fontsize
+            Label size of x- and y- axis ticks, by default None.
+            If None then using global fontsize
         **kwargs : kwargs
-            kwargs passed to `matplotlib.axes.Axes.set_xlabel()`
+            Keyword arguments passed to `matplotlib.axes.Axes.set_xlabel()`
         """
         labelsize = self.fontsize if labelsize is None else labelsize
         self.axis_top.tick_params(axis="y", labelsize=labelsize, **kwargs)
@@ -531,11 +551,11 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         xmin : float, optional
-            min of x-axis, by default None
+            Min of x-axis, by default None
         xmax : float, optional
-            max of x-axis, by default None
+            Max of x-axis, by default None
         **kwargs : kwargs
-            kwargs passed to `matplotlib.axes.Axes.set_xlim()`
+            Keyword arguments passed to `matplotlib.axes.Axes.set_xlim()`
         """
         self.axis_top.set_xlim(
             self.xmin if xmin is None else xmin,
@@ -555,13 +575,13 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         plot_name : str
-            file name of the plot
+            File name of the plot
         transparent : bool, optional
-            if plot transparent, by default True
+            If plot transparent, by default False
         dpi : int, optional
-            dpi for plotting, by default 400
+            DPI for plotting, by default 400
         **kwargs : kwargs
-            kwargs passed to `matplotlib.figure.Figure.savefig()`
+            Keyword arguments passed to `matplotlib.figure.Figure.savefig()`
         """
         logger.debug("Saving plot to %s", plot_name)
         self.fig.savefig(
@@ -577,7 +597,7 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         **kwargs: kwargs
-            kwargs from `matplotlib.figure.Figure.tight_layout()`
+            Keyword arguments from `matplotlib.figure.Figure.tight_layout()`
         """
         self.fig.tight_layout(**kwargs)
 
@@ -645,14 +665,14 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         handles :  list
-            list of matplotlib.lines.Line2D object returned when plotting
-        ax_mpl : axis
-            matplotlib.axis object where the legend should be plotted
+            List of matplotlib.lines.Line2D object returned when plotting
+        ax_mpl : matplotlib.axis.Axes
+            `matplotlib.axis.Axes` object where the legend should be plotted
         labels : list, optional
-            plot labels. If None, the labels are extracted from the `handles`.
+            Plot labels. If None, the labels are extracted from the `handles`.
             By default None
         **kwargs : kwargs
-            kwargs which can be passed to matplotlib axis
+            Keyword arguments which can be passed to matplotlib axis
         """
         ax_mpl.legend(
             handles=handles,
@@ -671,14 +691,14 @@ class PlotBase(PlotObject):
         Parameters
         ----------
         ratio_panel : int
-            ratio panel either 1 or 2
+            Indicates which ratio panel to modify (either 1 or 2).
         label : str
             y-axis label of the ratio panel
 
         Raises
         ------
         ValueError
-            if requested ratio panels and given ratio_panel do not match.
+            If requested ratio panels and given ratio_panel do not match.
         """
         # TODO: could add possibility to specify ratio label as function of rej_class
         if self.n_ratio_panels < ratio_panel and ratio_panel not in [1, 2]:
@@ -693,7 +713,7 @@ class PlotBase(PlotObject):
     def initialise_plot(self):
         """Calls other methods which are usually used when plotting"""
         self.set_title()
-        self.set_logy()
+        self.set_log()
         self.set_y_lim()
         self.set_xlabel()
         self.set_ylabel(self.axis_top)